Purge build volume files rather than re-creating it
If vscode has created a container then deleting the build volume will because the volume is in use. Deleting the vscode container is undesirable as vscode installs extensions. etc in this container and they can take a while to download.

This PR removes the "delete build volume on clean" behaviour and replaces it with a "delete all files/folder in build volume" behaviour.