Run voice2json using python -m voice2json
This build allows you to run voice2json as a python module. Note that
the paths might not be setup correctly as the official way to run voice2json
is by using the bin/voice2json script, which sets various variables.

This however is a significant milestone because we can at least run the command
and have a reproducible build in a container.