Build stages in the Dockerfile using parallel jobs.
The Dockerfile has been split into multiple stages:
- `llvm`
- `intel-graphics-compiler-and-compute-runtime`
- `openblas`
- `other-dependencies`
- `final-image`

The last one, `final-image` is the name of the stage which copies all
the files together from the other stages and produces a working
container. Images are cached on hub.docker.com.

Note: Reusable workflows are locked to a particular commit instead of
HEAD, this is a bug in github workflows and will most likely be fixed
soon.