Fix output buffer corruption bug in the stupidest possible way.
Corruption would fix on a screen redraw (e.g. window resize or pressing R), so on ERR input (i.e. no input) we redraw the screen, along with a 100ms wait to make sure we aren't spam redrawing the whole screen. Downside: I think the 100ms wait is blocking (but only when we don't have any input to process), Upside: it fixes the character corruption bug. This is dumb af but so far it seems to fix things and scripttuner is lightweight enough that imo we can cop 100ms blocking lol.